pyproject.toml
[tool.poetry]
name = "pet autotests project"
version = "1.0"
description = "Pet project for training and fun"
authors = ["Alexander Belozerov <asbelozerov@sbercloud.ru>"]
license = "FREE"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
pytest = "^7.1.2"
allure-pytest = "^2.9.45"
colorlog = "^6.6.0"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
